@@ -1,95 +1,61 @@
 package bnet
 
 import (
-	"crypto/tls"
+	"context"
+	"fmt"
+	"strings"
 
-	p "github.com/Gophercraft/core/bnet/bgs/protocol"
-	cn "github.com/Gophercraft/core/bnet/bgs/protocol/connection/v1"
-	"github.com/Gophercraft/log"
-	"github.com/golang/protobuf/proto"
-	"github.com/superp00t/etc"
+	"github.com/Gophercraft/core/bnet/rpc"
+	"github.com/Gophercraft/core/bnet/util"
 )
 
-func Dial(endpoint string) (*Conn, error) {
-	cn := new(Conn)
-	var err error
-	cn.c, err = tls.Dial("tcp", endpoint, &tls.Config{
-		InsecureSkipVerify: true,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return cn, nil
+type Client struct {
+	dial_options []DialOption
+	connection   *Connection
+	// registered services
+	services map[util.ServiceHash]*registered_service
 }
 
-func (c *Conn) Emit(h *p.Header, data proto.Message) {
-	body, err := proto.Marshal(data)
-	if err != nil {
-		panic(err)
-	}
+type ClientOption func(*Client)
 
-	sz := uint32(len(body))
-	h.Size = &sz
-
-	head, err := proto.Marshal(h)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	log.Dump("body", body)
-	log.Dump("head", head)
-
-	pkt := etc.NewBuffer()
-	pkt.WriteBigUint16(uint16(len(head)))
-	_, err = c.c.Write(pkt.Bytes())
-	if err != nil {
-		panic(err)
-	}
-
-	_, err = c.c.Write(head)
-	if err != nil {
-		panic(err)
-	}
-
-	_, err = c.c.Write(body)
-	if err != nil {
-		panic(err)
-	}
+func NewClient(dial_options ...DialOption) (client *Client) {
+	client = new(Client)
+	client.dial_options = dial_options
+	client.services = make(map[util.ServiceHash]*registered_service)
+	return
 }
 
-func (c *Conn) Connect() error {
-	tr := true
-	cr := &cn.ConnectRequest{
-		UseBindlessRpc: &tr,
+func (client *Client) RegisterService(service_desc *rpc.ServiceDesc, service any) (err error) {
+	client.services[service_desc.ServiceHash] = &registered_service{
+		descriptor: service_desc,
+		service:    service,
 	}
+	return
+}
 
-	methodID := uint32(1)
-	serviceID := uint32(0)
-	token := uint32(0)
+func (client *Client) get_registered_service(hash util.ServiceHash) (service *registered_service) {
+	service = client.services[hash]
+	return
+}
 
-	head := &p.Header{
-		ServiceId: &serviceID,
-		MethodId:  &methodID,
-		Token:     &token,
+func (client *Client) Connect(portal string) (err error) {
+	address := portal
+	if !strings.Contains(address, ":") {
+		address += ":1119"
 	}
-
-	c.Emit(head, cr)
-
-	resph, data, err := c.ReadHeader()
+	fmt.Println("Dialing", address)
+	client.connection, err = DialContext(context.WithValue(context.TODO(), client_context_key, client), address, client.dial_options...)
 	if err != nil {
-		log.Warn("Err reading header", err)
-		return err
+		return
 	}
 
-	rsp := new(cn.ConnectResponse)
-	err = proto.Unmarshal(data, rsp)
-	if err != nil {
-		return err
-	}
+	return
+}
 
-	log.Dump("resph", resph)
-	log.Dump("rsp", rsp)
+func (client *Client) Disconnect() (err error) {
+	return client.connection.Close()
+}
 
-	return nil
+func (client *Client) Connection() *Connection {
+	return client.connection
 }
