@@ -1,9 +1,13 @@
 package main
 
 import (
+	"fmt"
+	"os"
+	"path/filepath"
+
 	"github.com/Gophercraft/core/realm/wdb/models"
 	"github.com/Gophercraft/core/tempest"
-	"github.com/Gophercraft/text"
+	"github.com/Gophercraft/log"
 )
 
 type PlayerCreateInfo struct {
@@ -21,35 +25,143 @@ func (pci PlayerCreateInfo) TableName() string {
 	return "Playercreateinfo"
 }
 
-func extractPlayerCreateInfo() {
-	fl := openFile("DB/PlayerCreateInfo.txt")
-	wr := text.NewEncoder(fl)
+func extractPlayerCreateInfoRaceClass(race models.Race, class models.Class) {
+	var pCreateInfo PlayerCreateInfo
+	found, err := DB.Where("race = ?", race).Where("class = ?", class).Get(&pCreateInfo)
+	if err != nil {
+		log.Warn(err)
+		return
+	}
 
-	var pCreateInfo []PlayerCreateInfo
+	if !found {
+		log.Warn("Could not find race class combo for", raceName[race], className[class])
+		return
+	}
 
-	err := DB.Find(&pCreateInfo)
+	createInfo := new(models.PlayerCreateInfo)
+	createInfo.Race = race
+	createInfo.Class = class
+	createInfo.Placement.Map = pCreateInfo.Map
+	createInfo.Placement.Zone = pCreateInfo.Zone
+	createInfo.Placement.Position = tempest.C4Vector{
+		pCreateInfo.X,
+		pCreateInfo.Y,
+		pCreateInfo.Z,
+		pCreateInfo.O,
+	}
+
+	// Get items
+	// var pCreateItems []PlayerCreateItem
+	// err = DB.Where("race = ?", race).Where("class = ?", class).Find(&pCreateItems)
+	// if err != nil {
+	// 	panic(err)
+	// }
+
+	// for _, pCreateItem := range pCreateItems {
+	// 	itemTemplate, ok := allItems[fmt.Sprintf("it:%d", pCreateItem.ItemID)]
+	// 	if !ok {
+	// 		panic(pCreateItem.ItemID)
+	// 	}
+
+	// 	var item models.PlayerCreateItem
+	// 	switch itemTemplate.InventoryType {
+	// 	case models.IT_Unequippable:
+	// 		item.Equip = models.EquipInventory
+	// 	case models.IT_Bag:
+	// 		item.Equip = models.EquipContainer
+	// 	default:
+	// 		item.Equip = models.EquipPaperDoll
+	// 	}
+
+	// 	item.Item = itemTemplate.ID
+	// 	item.Amount = pCreateItem.Amount
+
+	// 	createInfo.Items = append(createInfo.Items, item)
+	// }
+
+	var pCreateSpells []PlayerCreateInfoSpell
+	err = DB.Where("race = ?", race).Where("class = ?", class).Find(&pCreateSpells)
 	if err != nil {
 		panic(err)
 	}
 
-	for _, pci := range pCreateInfo {
-		mod := &models.PlayerCreateInfo{
-			Race:  models.Race(pci.Race),
-			Class: models.Class(pci.Class),
-			Position: tempest.C4Vector{
-				pci.X,
-				pci.Y,
-				pci.Z,
-				pci.O,
-			},
-			Map:  pci.Map,
-			Zone: pci.Zone,
+	for _, pCreateSpell := range pCreateSpells {
+		createInfo.Abilities = append(createInfo.Abilities, models.PlayerCreateAbility{
+			Spell: pCreateSpell.Spell,
+			Note:  pCreateSpell.Note,
+		})
+	}
+
+	var pCreateActionButtons []PlayerCreateActionButton
+	err = DB.Where("race = ?", race).Where("class = ?", class).Find(&pCreateActionButtons)
+	if err != nil {
+		panic(err)
+	}
+
+	for _, pCreateActionButton := range pCreateActionButtons {
+		actionbutton := models.PlayerCreateActionButton{
+			Button: pCreateActionButton.Button,
+			Action: pCreateActionButton.Action,
 		}
 
-		if err := wr.Encode(mod); err != nil {
-			panic(err)
+		switch pCreateActionButton.Type {
+		case 0:
+			actionbutton.Type = models.ActionSpell
+		case 64:
+			actionbutton.Type = models.ActionMacro
+		case 128:
+			actionbutton.Type = models.ActionItem
+		default:
+			panic(pCreateActionButton.Type)
+		}
+
+		createInfo.ActionButtons = append(createInfo.ActionButtons, actionbutton)
+	}
+
+	file := openTextFile(fmt.Sprintf("DB/PlayerCreateInfo/%s_%s.txt", raceName[race], className[class]))
+	if err := file.Encode(createInfo); err != nil {
+		panic(err)
+	}
+	file.close()
+}
+
+func extractPlayerCreateInfo() {
+	os.MkdirAll(filepath.Join("DB", "PlayerCreateInfo"), 0700)
+
+	for _, race := range races {
+		for _, class := range classes {
+			extractPlayerCreateInfoRaceClass(race, class)
 		}
 	}
 
-	fl.Close()
+	// fl := openFile("DB/PlayerCreateInfo.txt")
+	// wr := text.NewEncoder(fl)
+
+	// var pCreateInfo []PlayerCreateInfo
+
+	// err := DB.Find(&pCreateInfo)
+	// if err != nil {
+	// 	panic(err)
+	// }
+
+	// for _, pci := range pCreateInfo {
+	// 	mod := &models.PlayerCreateInfo{
+	// 		Race:  models.Race(pci.Race),
+	// 		Class: models.Class(pci.Class),
+	// 		Position: tempest.C4Vector{
+	// 			pci.X,
+	// 			pci.Y,
+	// 			pci.Z,
+	// 			pci.O,
+	// 		},
+	// 		Map:  pci.Map,
+	// 		Zone: pci.Zone,
+	// 	}
+
+	// 	if err := wr.Encode(mod); err != nil {
+	// 		panic(err)
+	// 	}
+	// }
+
+	// fl.Close()
 }
