@@ -9,293 +9,95 @@
 package datapack
 
 import (
-	"archive/zip"
-	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
-	"path/filepath"
-	"sort"
-	"strings"
-	"time"
 
 	"github.com/Gophercraft/text"
-	"github.com/superp00t/etc"
 )
 
-type File io.ReadCloser
-
-type WriteFile io.WriteCloser
-
-// Driver describes a mechanism for loading a datapack.
-type Driver interface {
-	Init(at string) (Opts, error)
-	ReadFile(at string) (File, error)
-	WriteFile(at string) (WriteFile, error)
-	List() []string
-	Close() error
+type PackDependency struct {
+	// The ID (found in PackInfo) to depend on
+	ID string
+	// The minimum acceptable version of this we depend on
+	MinimumVersion uint32
 }
 
-type Opts uint64
-
-const None Opts = 0
-
-const (
-	Read Opts = 1 << iota
-	Write
-)
-
-var (
-	drivers = map[string]func() Driver{
-		"flatfile": func() Driver {
-			return new(flatFile)
-		},
-
-		"archive": func() Driver {
-			return new(archive)
-		},
-	}
-)
-
-func RegisterDriver(key string, value func() Driver) {
-	if drivers[key] != nil {
-		panic("datapack: " + key + " already registered")
-	}
-
-	drivers[key] = value
-}
-
-type PackConfig struct {
-	Name           string
-	Description    string
-	Author         string
-	Version        string
+type PackInfo struct {
+	// the datapack's internal name "Gophercraft/example"
+	ID string
+	// The datapack's display name "Gophercraft Example Pack"
+	Name string
+	// Incremental version number of the pack itself
+	Version uint32
+	// SemVer of the required minimum Gophercraft Core version (x.x.x)
+	MinimumCoreVersion string
+	// The description of the datapack
+	Description string
+	// URL to the Git repository of the datapack "https://github.com/Gophercraft/example.git"
+	Repository string
+	//
+	Authors      []string
+	Base         bool
+	Dependencies []PackDependency
+	// Database tables to be overwritten
 	OverrideTables []string
-	Depends        []string
-	ServerScripts  []string
-	ClientScripts  []string
 }
 
 type Pack struct {
-	PackConfig
-	Opts
-	Driver
+	info                    PackInfo
+	reader                  reader
+	dependencies            []*Pack
+	transitive_dependencies []*Pack
 }
 
-func (p *Pack) WriteBytes(path string, data []byte) error {
-	file, err := p.WriteFile(path)
-	if err != nil {
-		return err
-	}
-
-	_, err = file.Write(data)
-	if err != nil {
-		return err
-	}
-
-	err = file.Close()
-	return err
+func (pack *Pack) Open(path string) (file io.ReadCloser, err error) {
+	return pack.reader.Open(path)
 }
 
-func (p *Pack) ReadBytes(path string) ([]byte, error) {
-	file, err := p.Driver.ReadFile(path)
-	if err != nil {
-		return nil, err
-	}
-
-	b, err := ioutil.ReadAll(file)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := file.Close(); err != nil {
-		return nil, err
-	}
-
-	return b, nil
+func (pack *Pack) List() (list []string) {
+	list = pack.reader.List()
+	return
 }
 
-func OpenPack(path string) (*Pack, error) {
-	p := new(Pack)
-	driv := ""
-	if etc.ParseSystemPath(path).IsDirectory() {
-		driv = "flatfile"
+func Open(path string) (pack *Pack, err error) {
+	var fi os.FileInfo
+	fi, err = os.Stat(path)
+	if err != nil {
+		return
+	}
+
+	var r reader
+
+	if fi.IsDir() {
+		r, err = open_directory(path)
 	} else {
-		driv = "archive"
+		r, err = open_archive(path)
 	}
-
-	var err error
-
-	p.Driver = drivers[driv]()
-	p.Opts, err = p.Driver.Init(path)
-	if err != nil {
-		return nil, err
-	}
-
-	yb, err := p.ReadBytes("Pack.txt")
-	if err != nil {
-		return nil, err
-	}
-
-	err = text.Unmarshal(yb, &p.PackConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return p, nil
-}
-
-func Open(directory string) (*Loader, error) {
-	bp := etc.ParseSystemPath(directory)
-
-	l := new(Loader)
-
-	packs, err := ioutil.ReadDir(directory)
-	if err != nil {
-		return nil, err
-	}
-
-	var s []string
-
-	for _, v := range packs {
-		s = append(s, v.Name())
-	}
-
-	sort.Strings(s)
-
-	for _, v := range s {
-		path := bp.Concat(v).Render()
-		pack, err := OpenPack(path)
-		if err != nil {
-			return nil, err
-		}
-		l.Volumes = append(l.Volumes, pack)
-	}
-
-	return l, nil
-}
-
-func (p *Pack) FolderExists(path string) bool {
-	if !strings.HasSuffix(path, "/") {
-		path += "/"
-	}
-
-	list := p.List()
-	for _, v := range list {
-		if strings.HasPrefix(v, path) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (p *Pack) FolderList(path string) []string {
-	if !strings.HasSuffix(path, "/") {
-		path += "/"
-	}
-
-	var list []string
-
-	allFiles := p.List()
-	for _, file := range allFiles {
-		if strings.HasPrefix(file, path) {
-			list = append(list, file)
-		}
-	}
-
-	// Ensure consistent behavior
-	sort.Strings(list)
-
-	return list
-}
-
-func (p *Pack) Exists(path string) bool {
-	list := p.List()
-	for _, v := range list {
-		if v == path {
-			return true
-		}
-	}
-
-	return false
-}
-
-func AuthorDir(dir string, cfg PackConfig) (*Pack, error) {
-	if err := os.MkdirAll(dir, 0700); err != nil {
-		return nil, err
-	}
-
-	data, err := text.Marshal(cfg)
-	if err != nil {
-		panic(err)
-	}
-
-	err = ioutil.WriteFile(
-		filepath.Join(dir, "Pack.txt"),
-		data,
-		0700,
-	)
 
 	if err != nil {
-		return nil, err
+		return
 	}
 
-	pack, err := OpenPack(dir)
+	pack = new(Pack)
+	pack.reader = r
+
+	var info_file io.ReadCloser
+	info_file, err = pack.Open("Pack.txt")
 	if err != nil {
-		return nil, err
+		return
 	}
 
-	return pack, nil
-}
-
-func Author(cfg PackConfig) (*Pack, error) {
-	tempToken := etc.GenerateRandomUUID().String()
-	tempDir := etc.TmpDirectory().Concat(tempToken)
-
-	return AuthorDir(tempDir.Render(), cfg)
-}
-
-func (p *Pack) ZipToFile(filename string) error {
-	os.Remove(filename)
-
-	newZipFile, err := os.Create(filename)
+	decoder := text.NewDecoder(info_file)
+	err = decoder.Decode(&pack.info)
 	if err != nil {
-		return err
+		return
 	}
-	defer newZipFile.Close()
 
-	zipWriter := zip.NewWriter(newZipFile)
-	defer zipWriter.Close()
-	// Add files to zip
-	for _, file := range p.List() {
-		rdr, err := p.ReadFile(file)
-		if err != nil {
-			fmt.Println(file)
-			panic(err)
-			return err
-		}
+	info_file.Close()
 
-		if err = addFileToZip(zipWriter, file, rdr); err != nil {
-			return err
-		}
-
-		rdr.Close()
-	}
-	return nil
+	return
 }
 
-func (p *Pack) Delete() error {
-	switch f := p.Driver.(type) {
-	case *flatFile:
-		return os.RemoveAll(f.Base.Render())
-	case *archive:
-		return os.Remove(f.Path)
-	default:
-		return fmt.Errorf("unknown pack type")
-	}
-}
-
-func Timestamp() string {
-	return time.Now().Format("Mon Jan 2 15:04:05 MST 2006")
+func (pack *Pack) Info() (pack_info *PackInfo) {
+	return &pack.info
 }
