@@ -1,7 +1,7 @@
 // Package replay will provide utilities for recording and playing back game logs
 package replay
 
-import "github.com/Gophercraft/core/vsn"
+import "github.com/Gophercraft/core/version"
 
 type Version uint8
 
@@ -11,6 +11,6 @@ const (
 
 type Header struct {
 	Version
-	Build vsn.Build
+	Build version.Build
 	Map   uint32
 }
