@@ -5,10 +5,9 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/Gophercraft/core/app/config"
 	"github.com/Gophercraft/core/format/content"
-	"github.com/Gophercraft/core/home/config"
-	"github.com/Gophercraft/core/home/dbsupport"
-	"github.com/Gophercraft/core/vsn/detection"
+	"github.com/Gophercraft/core/version/detection"
 )
 
 type WorldOptions struct {
@@ -20,7 +19,6 @@ type WorldConfigurator struct {
 	Config       *config.World
 	GamePath     string
 	ConfigName   string
-	Registrar    Registrar
 }
 
 func (co *Configurator) NewWorldConfigurator() *WorldConfigurator {
@@ -30,45 +28,31 @@ func (co *Configurator) NewWorldConfigurator() *WorldConfigurator {
 	return wc
 }
 
-func (w *WorldConfigurator) LoadConfig(confName string) (err error) {
-	w.Config, err = config.LoadWorld(filepath.Join(w.Configurator.Dir, confName))
-	if err != nil {
-		return
-	}
-
-	return
-}
-
-func (w *WorldConfigurator) SetConfigName(confName string) {
-	w.Config.Dir = filepath.Join(w.Configurator.Dir, confName)
-	w.ConfigName = confName
-}
-
 func (w *WorldConfigurator) ConfigExists() bool {
-	_, err := os.Stat(w.Config.Dir)
+	_, err := os.Stat(w.Config.Directory)
 	return err == nil
 }
 
 func (w *WorldConfigurator) RemoveConfigDir() error {
-	if w.Config.Dir == "" {
+	if w.Config.Directory == "" {
 		return fmt.Errorf("wizard: can't remove empty directory")
 	}
 
-	return os.RemoveAll(w.Config.Dir)
+	return os.RemoveAll(w.Config.Directory)
 }
 
 func (w *WorldConfigurator) GenerateConfigDir() error {
-	if err := os.MkdirAll(w.Config.Dir, 0700); err != nil {
+	if err := os.MkdirAll(w.Config.Directory, 0700); err != nil {
 		return err
 	}
 
 	if err := os.MkdirAll(
-		filepath.Join(w.Config.Dir, "Datapacks"),
+		filepath.Join(w.Config.Directory, "Datapacks"),
 		0700); err != nil {
 		return err
 	}
 
-	if err := w.Config.GenerateKeyPair(); err != nil {
+	if err := config.GenerateKeyPair(w.Config.Directory); err != nil {
 		return err
 	}
 
@@ -78,16 +62,16 @@ func (w *WorldConfigurator) GenerateConfigDir() error {
 func (wc *WorldConfigurator) SetVolume(path string) error {
 	var err error
 	wc.GamePath = path
-	wc.Config.Version, err = detection.DetectGame(wc.GamePath)
+	wc.Config.Build, err = detection.DetectGame(wc.GamePath)
 	return err
 }
 
 func (wc *WorldConfigurator) SetName(name string) {
-	wc.Config.RealmName = name
+	wc.Config.LongName = name
 }
 
 func (wc *WorldConfigurator) SetRealmType(t config.RealmType) error {
-	wc.Config.RealmType = t
+	wc.Config.Type = t
 	return nil
 }
 
@@ -99,12 +83,13 @@ func (wc *WorldConfigurator) SetRealmType(t config.RealmType) error {
 // }
 
 func (wc *WorldConfigurator) CreateDB() error {
-	return dbsupport.Create(wc.Config.DBDriver, wc.Config.DBURL)
+	return nil
+	// return dbsupport.Create(wc.Config.DBDriver, &wc.Config.DBURL)
 }
 
 func (w *WorldConfigurator) NewExtractor() (*Extractor, error) {
 	ex := new(Extractor)
-	ex.Dir = filepath.Join(w.Config.Dir, "Datapacks")
+	ex.Dir = filepath.Join(w.Config.Directory, "Datapacks")
 	var err error
 	ex.Source, err = content.Open(w.GamePath)
 	if err != nil {
