@@ -4,6 +4,7 @@ import (
 	"path/filepath"
 
 	"github.com/Gophercraft/core/datapack"
+	"github.com/Gophercraft/core/version"
 )
 
 const MapPackName = "!maps.zip"
@@ -16,11 +17,12 @@ func (ex *Extractor) ExtractMaps() error {
 	}
 
 	pack, err := ex.AuthorPack(tempPackDir, datapack.PackConfig{
-		Name:        "Gophercraft Base Maps",
-		Author:      exAuthor,
-		Description: "The base map data required by Gophercraft Core." + ex.generationNotice(),
-		Version:     "1.0",
-		Depends:     ex.dependencies(),
+		Name:           "Gophercraft Base Maps",
+		Author:         exAuthor,
+		Description:    "The base map data required by Gophercraft Core." + ex.generationNotice(),
+		PackVersion:    "1.0",
+		MinCoreVersion: version.GophercraftVersion.String(),
+		Depends:        ex.dependencies(),
 	})
 
 	if err != nil {
